Update `BloodSugarMeasurements` DB schema
- Change `BloodSugarReading#value` type from `Int` to `Float`

+++ app/src/androidTest/java/org/simple/clinic/TestData.kt
@@ -849,7 +849,7 @@ class TestData @Inject constructor(
 
   fun bloodSugarMeasurement(
       uuid: UUID = UUID.randomUUID(),
-      reading: BloodSugarReading = BloodSugarReading(faker.number.between(30, 1000), Random),
+      reading: BloodSugarReading = BloodSugarReading(faker.number.between(30, 1000).toFloat(), Random),
       patientUuid: UUID = UUID.randomUUID(),
       recordedAt: Instant = Instant.now(userClock),
       userUuid: UUID = qaUser().uuid,
